Updated docs for timeout to be float instead of int

+++ twilio/base/domain.py
@@ -30,7 +30,7 @@ class Domain(object):
         :param object data: The request body.
         :param dict headers: The HTTP headers.
         :param tuple auth: Basic auth tuple of (username, password)
-        :param int timeout: The request timeout.
+        :param float timeout: The request timeout.
         :param bool allow_redirects: True if the client should follow HTTP
         redirects.
         """
